The app now has a landing page which can navigate user to the page with our fetches. Once on the page with our fetches you can click the title in the navbar to take you back to home
