+++ react-app/src/App.scss
@@ -9,15 +9,22 @@
   box-shadow: 5px 5px 10px;
   
   
-      a {
+      h1{
         display: flex;
         margin: auto;
         letter-spacing: 20px;
         font-family: 'Fascinate', sans-serif;
         color: #f6f2dd;
         font-size: xx-large;
-       
-       
+        
+          a{
+            color: #f6f2dd;
+            
+              &:hover{
+                text-decoration: none;
+                transform: scale(1.1);
+            }
+          }
       }
   }
