--- react-app/src/components/apps/Nasa api/styles.scss
@@ -1,95 +1,0 @@
-body {
-    background-color: #f6f2dd;
-}
-
-.main {   
-   width: auto;
-   float: left;
-   padding: 50px;
-   margin-top: 15vh;
-   margin-left: 20px;
-   font-family: 'Inter', sans-serif;
-   background-color:  rgba(201, 108, 68, 0.6);
-   border-radius: 20px;
-}
-
-.textDiv{
-    display: flex;
-    flex-direction: column;
-    text-align: center;
-    background-color:  #f6f2dd;
-    border-radius: 20px;
-    margin: auto;
-    width: 50%;
-    height: auto;
-    box-shadow: 5px 5px 10px;
-
-    h3{
-        font-weight: bolder;
-        margin-top: 10px;
-    }
-    p {
-        display: flex;
-        font-size:x-large;
-        align-self: center;
-        padding: 10px;
-    }
-}
-
-.imgDiv{
-    display: flex;
-    flex-direction: column;
-    justify-content: flex-end;
-    background-color:  #f6f2dd;
-    border-radius: 20px;
-    //width: 50%;
-    margin: auto;
-    margin-top: 20px;
-    box-shadow: 5px 5px 10px;
-
-    img {
-        margin: auto;
-        height: 80%; 
-        width: 50%; 
-        margin-top: 20px ;
-        image-rendering: crisp-edges;
-        border-radius: 50%;
-        box-shadow: 5px 5px 10px;
-        margin-bottom: 20px;
-        
-        
-    }
-}
-
-button{
-    margin: auto;
-    margin-top: 10px;
-    margin-bottom: 10px;
-    background-color:rgba(68, 111, 201, 0.5) ;
-    color:  #f6f2dd;
-    font-size:x-large;
-    border-radius: 8px;
-    padding:20px; 
-    border:none;
-    box-shadow: 2px 2px 5px black;
-    font-family: 'Fascinate';
-
-    &:hover{
-        transform: scale(1.01);
-        transition-duration: 0.3s;
-        box-shadow: 3px 3px 5px black;
-     
-    }
-}
-
-.weatherContainer{
-    display: flex;
-    justify-content: center;
-    flex-direction: column;
-    width: 20%;
-
-    button{
-    margin-top: 15vh;
-
-    }
-}
