+++ react-app/src/components/site/footer.scss
@@ -1,6 +1,6 @@
 footer{
     display: flex;
-    margin-top: 100vh;
-    justify-content: center;
+    justify-content:center;
+    margin-top: 10vh;
     align-content: center;
 }
