+++ react-app/src/components/site/home.scss
@@ -1,0 +1,52 @@
+
+
+.landingDiv {
+    margin: auto;
+    margin-top: 5%;
+    //border: 1px solid black;
+    border-radius: 50%;
+    box-shadow: 5px 5px 10px black;
+    width:max-content;
+    height: max-content;
+    background-color: rgba(201, 108, 68, 0.6);
+
+    .landing{
+    
+        display: flex;
+        align-content: center;
+        justify-content: center;
+        flex-direction: row;
+        width:70vh;
+        height:70vh;
+    
+        h1{
+            display: flex;
+            align-content: center;
+            padding-bottom: 5px;
+            margin-top: 50%;
+            margin-bottom: 10%;
+            font-family: 'Fascinate';
+            color: #f6f2dd;
+            letter-spacing: 2px;
+            border-bottom: 5px ridge #f6f2dd;
+            border-radius: 4px;
+            text-shadow: 2px 2px 3px black;
+
+        }
+
+        button{
+            display: flex;
+            padding: 0%;
+            text-align: center;
+            align-self: baseline;
+            background-color:rgba(68, 111, 201, 0.8);
+        }
+
+        a {
+            padding: 20px;
+            color:#f6f2dd;
+            text-decoration: none;
+        }
+    }
+}
+
